Projects/AKS/requirements.txt
# 项目依赖包列表

# Web框架和服务器
fastapi==0.103.2
uvicorn==0.23.2
python-multipart==0.0.6

# 数据库驱动
psycopg2-binary==2.9.9

# 数据处理和验证
pydantic==2.4.2
python-dateutil==2.8.2

# 类型提示
typing-extensions==4.8.0

# 其他依赖
requests==2.31.0
python-dotenv==1.0.0
email-validator==2.1.0.post1

# 开发工具
pytest==7.4.3
black==23.10.1
flake8==6.1.0
mypy==1.6.1